requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
falcon==2.0.0
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-pubsub==0.40.0
googleapis-common-protos==1.5.10
grpc-google-iam-v1==0.11.4
grpcio==1.20.1
gunicorn==19.9.0
idna==2.8
mccabe==0.6.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pydocstyle==3.0.0
pyflakes==2.1.1
pylama==7.7.1
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
snowballstemmer==1.2.1
urllib3==1.24.3
